Create README.md
The used car market in India is a dynamic and ever-changing landscape. Prices can fluctuate wildly based on a variety of factors including the make and model of the car, its mileage, its condition and the current market conditions. As a result, it can be difficult for sellers to accurately price their cars.

This dataset contains information about used cars. This data can be used for a lot of purposes such as Used Car Price Prediction using different Machine Learning Techniques.

The goal of the project is to predict the price of a second hand car that will be listed in future in the E-Commerce website using statistical modelling and machine learning techniques.